requirements.txt
pytest~=8.3.5
selenium~=4.31.0
webdriver-manager
requests~=2.32.3
allure-pytest
faker~=37.4.0
flake8
allure-python-commons~=2.14.2